nanvm-lib: Number, the VM's number type, holds one NaN by construction
`Unpacked::Number`, `Numeric::Number` and `Primitive::Number` carried a
bare `f64`, so the one-`NaN` invariant #2152 gave the VM was a rule: a
canonicalizing constructor beside a public variant that skipped it, a
match in `Naive`'s `From` to catch what skipped, and a trait doc asking
every other VM to do the same.

`Number` makes it a fact of the type. It is a `#[repr(transparent)]`
newtype over `f64` with a private field: the only way in is `From<f64>`,
which canonicalizes every `NaN` to the quiet, positive, empty-payload one,
the way out is `From<Number> for f64`, and each arithmetic operator on it
passes its result back through the same conversion, since hardware leaves a
sign or payload on some `NaN`s — `-NaN`, x86's `0 / 0`. The three variants
carry a `Number`, `Dispatch::number` and `Any::to_number` speak it, and
`IVm` gains `From<Number>` so a `Number` is an `Any` like every other value.
`Unpacked::number`, the `Naive` match and the doc rule are gone: there is
nothing left to skip.

The `f64` helpers — `to_int32`, `canonical_index`, `number_to_string`,
the comparisons — keep their type and are called with an explicit
conversion where a variant is opened; moving number semantics onto the type
is a change of its own, not this rename's.

+++ nanvm-lib/src/vm/unstable/mod.rs
@@ -50,7 +50,7 @@ mod test {
     #[test]
     fn numbers_by_their_bits() {
         let bits = |v: Any<Naive>| match v.into() {
-            Unpacked::Number(x) => x.to_bits(),
+            Unpacked::Number(x) => f64::from(x).to_bits(),
             _ => panic!("a number"),
         };
         assert_eq!(bits(f64_any(0x4002666666666666)), 2.3f64.to_bits());
